daar_backend/requirements.txt
Django==5.1.4
numpy==1.26.4
networkx==3.4.2
scipy==1.15.0
psycopg2-binary==2.9.10
tqdm==4.67.1
docker==5.0.3